% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_helper_fxns.R
\name{compare_local_and_gdrive}
\alias{compare_local_and_gdrive}
\title{Compare the Local File and the Current Version of the Gdrive File}
\usage{
compare_local_and_gdrive(l_path, g_path)
}
\arguments{
\item{l_path}{the output of \code{parse_local_path()}, identifying the local file.}

\item{g_path}{the output of \code{parse_dribble()}, identifiying the Gdrive file}
}
\value{
Returns a list of checks, including whether the local and Gdrive files were identical, have the same size or
modification times, and if needed, the raw bytes of the Gdrive file.
}
\description{
A background helper function for \code{gdrive_upload()} and \code{gdrive_download()} that helps determine whether those
operations are necessary (e.g., skips these actions if the files are already identical).
}
