% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_dir.R
\name{gdrive_dir}
\alias{gdrive_dir}
\title{Show the Folder Structure of the Shared Google Drive}
\usage{
gdrive_dir(shared_id = "default_gdrive_id", folder = NULL)
}
\arguments{
\item{shared_id}{An alias of a Shared Google Drive. By default, searches for the \code{default_gdrive_id} variable in your
.Renviron file.}

\item{folder}{A dribble or a character file path of folders, i.e., "Projects", that will narrow down your search.}
}
\value{
Returns a data frame of the folder structure of the Shared Google Drive.
}
\description{
Prints the structure of the Shared Google Drive to the console. It provides a count of the files within
each folder and all subfolders. Items in the 'path' column can be copy-pasted to use as the
\code{gdrive_path} argument of the \code{gdrive_set_dribble(gdrive_path)} function. Highly recommended to use the \code{folder}
argument as it will speed up your search. Note that this function only works for shared drives in which you have
member-level access, not just folder-level access.
}
\details{
Use this function to retrieve the file paths of every folder in the Shared Google Drive.
Once you have found the folder you would like to upload to or download from, copy its \code{path} to use
as the \code{gdrive_path} argument of the \code{gdrive_set_dribble()} function.
}
\examples{
\dontrun{
   gdrive_dir()
}

}
\seealso{
\code{gdrive_set_dribble()}
}
