% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_helper_fxns.R
\name{gdrive_gcw_setup}
\alias{gdrive_gcw_setup}
\title{Setup your Google Shared Drive token on the GCW}
\usage{
gdrive_gcw_setup()
}
\value{
Opens a file explorer window that contains your token
}
\description{
Setup your Google Shared Drive token on the GCW
}
\details{
In order to access the Shared Google Drive in the cloud, you have to upload credentials from your local machine to
your workstation. This function helps you locate this file on your local machine.
}
\keyword{internal}
