% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_helper_fxns.R
\name{gdrive_get_shared_id}
\alias{gdrive_get_shared_id}
\title{Specify a Google Shared Drive}
\usage{
gdrive_get_shared_id(shared_id)
}
\arguments{
\item{shared_id}{Either an alias for a shared_id stored in your .Renviron file or the drive_id of Google Shared Drive
that you have access to.}
}
\value{
Returns the drive_id of a shared google drive
}
\description{
Specify a Google Shared Drive
}
\details{
By default, the shared_id argument of the gdrive_set_dribble() and gdrive_dir() functions check your .Renviron file
for a value named \code{default_gdrive_id} that contains the drive_id of a Google Shared Drive.
}
\keyword{internal}
