% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_keep_forever.R
\name{gdrive_keep_forever}
\alias{gdrive_keep_forever}
\title{Set all versions of a file's keepForever option to TRUE}
\usage{
gdrive_keep_forever(local_path, gdrive_dribble, ...)
}
\arguments{
\item{local_path}{the local path to the file you wish to download, including the directory (folder paths) and the name of the file, which should match the name of the file on the Gdrive.}

\item{gdrive_dribble}{the dribble class object representing the folder in the Gdrive where your desired file resides.}

\item{...}{not used, but allows you to easily replace calls to gdrive_versions without removing additional args.}
}
\value{
Returns a message whether updates to \code{keepForever} were successful.
}
\description{
If a file on the drive is ever manually uploaded (instead of using \code{gdrive_upload()}), it's \code{keepForever} option
will not be set to TRUE automatically. This will cause checks in the package to throw warnings. This function will
scan through all versions of a file where keepForever is FALSE and set them to TRUE.
}
