% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_ls.R
\name{gdrive_ls}
\alias{gdrive_ls}
\title{Show All Files in a Dribble}
\usage{
gdrive_ls(gdrive_dribble)
}
\arguments{
\item{gdrive_dribble}{a 1-row dribble class object created by \code{gdrive_set_dribble()}}
}
\value{
Prints a data.frame of the files contained in the specified Gdrive folder.

#' @seealso \code{gdrive_set_dribble()}
}
\description{
A reference function to print the names and metadata of files contained in a dribble.
}
\details{
If you do not know the folder path to the folder you want to interact with, run \code{gdrive_dir()} to get
the view the folder structure of the Share Google Drive and copy-paste the \code{path}
}
\examples{
\dontrun{
   gdrive_ls(gdrive_set_dribble(gdrive_path = 'Google Drive Test/'))
}

}
