% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_set_dribble.R
\name{gdrive_set_dribble}
\alias{gdrive_set_dribble}
\title{Designate a Gdrive Folder as a Target for Uploads and Downloads}
\usage{
gdrive_set_dribble(
  gdrive_path = NULL,
  shared_id = "default_gdrive_id",
  folder_id = NULL
)
}
\arguments{
\item{gdrive_path}{a character string of the folder path, with each folder name followed by \code{/}}

\item{shared_id}{a character string of the shared drive's ID or an alias for the Shared Drive to connect to. The
default, "default_gdrive_id", can be specified in your .Renviron file. See ?gdrive_get_shared_id for instructions
on how to set this up.}

\item{folder_id}{a character string of the folder's ID, which can be used instead of \code{gdrive_path} but must be
used for collaborators with only shared folder-level access to a shared drive's subfolder and not member access to
the shared drive.}
}
\value{
Returns a dribble class object
}
\description{
Given a folder path of the folder structure of the Shared Google Drive, this function will create a \code{dribble}
class object for use by key functions to interact with the drive, including \code{gdrive_upload()} and
\code{gdrive_download()}. By default, it searches the FMA Analytical Services Program shared drive.
}
\details{
If you do not know the folder path to the folder you want to interact with, run \code{gdrive_dir()} to get
the view the folder structure of the Share Google Drive and copy-paste the \code{path}.
}
\examples{
\dontrun{
   gdrive_set_dribble(gdrive_path = 'Google Drive Test/')
}

}
