% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_helper_fxns.R
\name{gdrive_token}
\alias{gdrive_token}
\title{For gdrive_ functions: check for googledrive token and if not active, use NOAA e-mail}
\usage{
gdrive_token()
}
\value{
Automatically uses the user's stored NOAA e-mail to connect if no token is already established
}
\description{
\code{drive_token()} is used by all \code{gdrive_} functions to ensure the user has authenticated their token to
communicate with the googledrive
}
\keyword{internal}
