% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_upload.R
\name{gdrive_upload}
\alias{gdrive_upload}
\title{Upload a File to the Google Shared Drive}
\usage{
gdrive_upload(local_path, gdrive_dribble, skip_prompt = FALSE)
}
\arguments{
\item{local_path}{the file path to the local file that you wish to upload or update.}

\item{gdrive_dribble}{the \code{dribble} object of the Gdrive folder you wish to upload to, created by \code{gdrive_set_dribble}.}

\item{skip_prompt}{skips the prompt asking the user if they are sure they want to upload when local is ahead of gdrive. FALSE by default.}
}
\value{
Returns messages regarding whether the upload was completed, and if so, the version number.
}
\description{
\code{gdrive_upload()} is used to either upload a new file to the specified Gdrive folder, or, if the file already exists,
updates the file by creating a new version. In the background, it assigns the modified dates of the local file to the
gdrive file/version, ensures older versions are kept forever, and checks to see whether uploading a new version is
necessary. Currently only works with \code{.rdata} and \code{.rds} type files. Nearly all other files we use should be
maintained the project's GitHub repository.
}
\examples{
\dontrun{
  gdrive_upload(local_path, gdrive_dribble)
}
}
