% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_versions.R
\name{gdrive_versions}
\alias{gdrive_versions}
\title{View the Revision History of a Data File on the Gdrive}
\usage{
gdrive_versions(local_path, gdrive_dribble)
}
\arguments{
\item{local_path}{the name of the .rdata or .rds file on the Gdrive}

\item{gdrive_dribble}{the \code{dribble} class object of the folder where your requested file resides.}
}
\value{
Returns a data.frame showing the file's revision history, including modification dates and users and file size
}
\description{
View the Revision History of a Data File on the Gdrive
}
\examples{
\dontrun{
  gdrive_versions(
     "geoff_mayhew_tutorial.rdata",
     gdrive_set_dribble("Google Drive Test/")
  )
}
}
