% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_helper_fxns.R
\name{parse_dribble}
\alias{parse_dribble}
\title{Find a File in the Gdrive Folder}
\usage{
parse_dribble(gdrive_dribble, l_path)
}
\arguments{
\item{gdrive_dribble}{the 1-row \code{dribble} class objet identifying the Gdrive folder to upload to or download from}

\item{l_path}{the output of \code{parse_local_path(local_path)}, which identifies the name of the file to search for within the \code{dribble}.}
}
\value{
Returns a list of \code{dribble} for the Gdrive file, a list of the file's revision history, and the file's current version number.
}
\description{
A background helper function for \code{gdrive_download()} and \code{gdrive_upload()} to identify the Gdrive files to interact with.
}
