% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_helper_fxns.R
\name{parse_local_path}
\alias{parse_local_path}
\title{Split a local path into the directory, filename, and extension}
\usage{
parse_local_path(local_path)
}
\arguments{
\item{local_path}{a character string of the local file path to the \code{.rdata} or \code{.rds} file to either upload or download.}
}
\value{
Returns a list of the parsed path as well as logical checks for whether the path was found or contains a version suffix.
}
\description{
A background helper function for \code{gdrive_download()} and \code{gdrive_upload()} to identify the local files to interact with.
}
\keyword{internal}
