% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_helper_fxns.R
\name{shared_drive_ls}
\alias{shared_drive_ls}
\title{\code{shared_dribe_ls()} is used in place of \code{googledrive::drive_ls()} which does not work for users with only
shared access to a subfolder of the shared drive and not root access.}
\usage{
shared_drive_ls(gdrive_dribble)
}
\arguments{
\item{gdrive_dribble}{a \code{dribble} containing the contents of a Shared Gdrive folder}
}
\value{
Returns a dribble of all items within the specified folder.
}
\description{
\code{shared_dribe_ls()} is used in place of \code{googledrive::drive_ls()} which does not work for users with only
shared access to a subfolder of the shared drive and not root access.
}
\keyword{internal}
